requirements.txt
Flask==2.3.2
python-whois==0.8.0
dnspython==2.4.2
